Linking sign-up process to firebase authentication
Changed the form to trigger a custom function that validates the password and then sends a POST request to the server. Server will then add the user into firebase authentication and also create a user profile in the firestore database. Upon success, it will send a status 200 in which the frontend will navigate to another page. If unsuccessful then a status code 500 will be sent instead